+++ packages/export/CHANGELOG.md
@@ -1,4 +1,15 @@
-﻿# 0.2.10
+﻿# 1.0.0
+* Strategies are now native JS async iterables (using async generators)
+* BREAKING:
+  * Most `exportStrategies` are removed - `bulk`, `page`, `stream`
+    * `bulk` can be replicated with `it-all`
+    * `stream` and `page` are handled by native `for await` loops
+  * `CSVStream` removed by outsourcing as much as possible to @fast-csv
+* Added `schemaToCSVTransforms` to make `fast-csv` schema aware
+* Added `andGroup` util
+
+
+# 0.2.10
 * `CSVStream` exportStrategy: document `omitFieldsFromResult`, refactoring.
 
 # 0.2.9